--- a/include/triton/Conversion/TritonGPUToLLVM/TargetInfoBase.h
+++ b/include/triton/Conversion/TritonGPUToLLVM/TargetInfoBase.h
@@ -94,7 +94,8 @@ public:
 
   virtual bool supportVectorizedAtomics() const = 0;
 
-  virtual bool supportLdStMatrix() const = 0;
+  virtual bool supportLdMatrix() const { return false; }
+  virtual bool supportStMatrix() const { return false; }
 
   // Annotate target specific information to local store operations during
   // lowering to LLVM.