… third_party/amd/lib/TritonAMDGPUToLLVM/TargetInfo.h
… third_party/nvidia/include/Dialect/NVGPU/IR/NVGPUOps.td
… third_party/nvidia/lib/NVGPUToLLVM/NVGPUToLLVMPass.cpp
… third_party/nvidia/lib/TritonNVIDIAGPUToLLVM/MemoryOpToLLVM.cpp
… third_party/nvidia/lib/TritonNVIDIAGPUToLLVM/TargetInfo.h