We combine the ldmatrix and stmatrix lowering paths, removing all the
previous limitations of the ldmatrix lowering.