@lezcano 2025-05-24
I run your script on this PR and on main and I got
```
pr
8.266822620820324 GB/s
1741.5311784108646 GB/s
1785.7254668400064 GB/s
1775.51729839168 GB/s
1786.7881878081714 GB/s

main
7.55253558571816 GB/s
1086.2517190956335 GB/s
1108.578369814276 GB/s
1105.584786392659 GB/s
1109.2609919631764 GB/s
```
so it looks like this PR makes this kernel faster?

After looking at the IR, I see that both the `local_load`s can be lowered to ldmatrix rather than unvectorised loads, so that should be better.
I think that the unswizzled layout will indeed cause bank conflicts in the read, but that was already the case in main, so nothing has changed there