@saagarjha 2025-05-25
Yes, your change definitely improves things and I'm also seeing the ldmatrix be emitted (also some nops, not entirely sure what that's about). I just wanted to point out that the lowering has bank conflicts; I wasn't sure if you intended to address that here too or not. FWIW here's the performance I'm seeing (on a GH200):

main (without your PR):
2.2134408831797434 GB/s
1365.554768760005 GB/s
1384.0195535865075 GB/s
1383.5943555669726 GB/s
1382.957048171502 GB/s

With your PR:
4.510313228611285 GB/s
2129.865040137383 GB/s
2131.3770124801213 GB/s
2150.20273448102 GB/s
2140.4941194726694 GB/s

With #6819:
3.07075893992839 GB/s
3327.373200864792 GB/s
3410.5260336012843 GB/s
3458.9858889174316 GB/s
3466.9743089842154 GB/s

So your change is a good 50% performance uplift or so, but the conflicts are costing it another 50% on top of that. The last one is not bottlenecked on shared memory requests at all but rather global memory bandwidth.

I guess I saw "ldmatrix" in your PR title and got really excited that I could drop my own patches, but it seems like that's not the case (yet?). It's totally reasonable if you didn't plan to fix the bank conflicts in this PR. I was just wondering if it is expected that the testcase I have should perform as it does with the changes here.